requirements.txt
protobuf
py-multibase
py-multihash
